Flask app is setup and running on my silk v3 space. Still need to test it using actual data from Cloudloop and on the Plaid silk account.
