+++ .gitignore
@@ -1,3 +1,4 @@
 .DS_Store
 .vscode/
-.DS_Store
+__pycache__/
+.env/
