+++ UVM-NRT-RoS/rockblock/flask/constants.py
@@ -1,0 +1,5 @@
+import os
+
+BASE_DIRECTORY = os.path.dirname(os.path.realpath(__file__))
+DATABASE_FILE = "database.db"
+DATABASE_URI = "sqlite:////" + os.path.join(BASE_DIRECTORY, DATABASE_FILE)
