+++ UVM-NRT-RoS/rockblock/flask/requirements.txt
@@ -1,7 +1,11 @@
 blinker==1.8.2
 click==8.1.7
 Flask==3.0.3
+Flask-SQLAlchemy==3.1.1
+greenlet==3.1.1
 itsdangerous==2.2.0
 Jinja2==3.1.4
-MarkupSafe==2.1.5
+MarkupSafe==3.0.2
+SQLAlchemy==2.0.36
+typing_extensions==4.12.2
 Werkzeug==3.0.4
