+++ UVM-NRT-RoS/rockblock/flask/wsgi.py
@@ -1,3 +1,3 @@
+import os
 import sys
-sys.path.insert(0, "/users/j/b/jbourde2/www-root/ciroh/api/v1")
 from app import app as application
